requirements.txt
PyQt6
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
selenium
openai-whisper
langchain
langchain-community
langchain-openai
webdriver-manager 